arm: sancloud_bbe_defconfig: Review and update
There were a whole bunch of options enabled in the TI staging kernel config for
am335x devices which we were missing. We also had a few options that were set
for the 4.14.y kernel and are now incorrect for the 4.9.y kernel.

Also, we need to set SMACK as the default and only Linux Security Module as well
as enable USB HID support.

After making these changes the config was regenerated via 'make savedefconfig'.

Signed-off-by: Paul Barker <paul@betafive.co.uk>